LPS-78224 Split portal-configuration-upgrade-util into api and impl
