+++ modules/apps/foundation/portal-configuration/portal-configuration-upgrade-api/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Portal Configuration Upgrade API
+Bundle-SymbolicName: com.liferay.portal.configuration.upgrade.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.configuration.upgrade
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Configuration Framework
