+++ modules/apps/foundation/portal-configuration/portal-configuration-upgrade-api/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
+	provided group: "org.osgi", name: "org.osgi.service.cm", version: "1.5.0"
+}
