+++ modules/apps/foundation/portal-configuration/portal-configuration-upgrade-api/src/main/resources/com/liferay/portal/configuration/upgrade/packageinfo
