+++ modules/apps/foundation/portal-configuration/portal-configuration-upgrade-util/build.gradle
@@ -4,4 +4,6 @@ dependencies {
 	provided group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
 	provided group: "org.osgi", name: "org.osgi.service.cm", version: "1.5.0"
 	provided group: "org.osgi", name: "org.osgi.service.component", version: "1.3.0"
+	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	provided project(":apps:foundation:portal-configuration:portal-configuration-upgrade-api")
 }
